Reorder FE components
