+++ client/src/styles.css
@@ -1,4 +1,4 @@
-.appheader {
+/* .appheader {
   display: flex;
   flex-direction: column;
   align-items: left;
@@ -155,4 +155,4 @@
 
 .x-100 {
   --value: 100%;
-}
+} */
